Add a branch to keep Ruby < 2.6 working
Prior to 2.6, periods were literals here and not sub expressions, so 
they were just emitted by sorcerer as strings. This commit fixed 2.6 
while breaking the test suite for <2.6:

0e31068dbfdf2ba23b38b8c1d4e2e48a976d36ab

So I added this (sadly, very obscure and not-self-documenting) branch to
the code to deal with it. Hopefully if anyone finds this WTF they'll 
git blame me